Add tests

+++ src/main/java/com/springboot/recipestore/RecipeService.java
@@ -2,8 +2,10 @@ package com.springboot.recipestore;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class RecipeService {
@@ -21,21 +23,42 @@ public class RecipeService {
     }
 
     Recipe selectRecipeByName(String name){
-        return recipeInterface.getById(name);
+        return recipeInterface.findById(name).isPresent() ? recipeInterface.findById(name).get() : new Recipe();
     }
 
     String insertNewRecipe(Recipe recipe){
-        recipeInterface.save(recipe);
-        return "Insertion Successful";
+        try {
+            recipeInterface.save(recipe);
+            return "Insertion Successful";
+        }catch(HttpClientErrorException.BadRequest br){
+            return br.getMessage();
+        }
     }
 
-    void deleteRecipeByName(String name){
-        recipeInterface.deleteById(name);
+    String deleteRecipeByName(String name){
+        if(recipeInterface.existsById(name)){
+            recipeInterface.deleteById(name);
+            return "Deletion Successful";
+        }else{
+            return "No Name exists for " + name;
+        }
     }
 
-    void updateRecipeByName(String name, String[] ingredients){
-        recipeInterface.updateRecipeByName(name, ingredients);
+    void updateIngredientsByName(String name, String[] ingredients){
+        if(recipeInterface.existsById(name)){
+            recipeInterface.updateIngredientsByName(name, ingredients);
+        }
     }
 
+    void updateMethodByName(String name, String[] method){
+        if(recipeInterface.existsById(name)){
+            recipeInterface.updateMethodByName(name, method);
+        }
+    }
 
+    void updateRecipeByName(String name, Recipe recipe){
+        if(recipeInterface.existsById(name)){
+            recipeInterface.updateRecipeByName(name, recipe.getName(), recipe.getSize(), recipe.getCategory());
+        }
+    }
 }
